term: kan structure
