+++ base/src/main/java/org/aya/core/term/MatchTerm.java
@@ -3,10 +3,21 @@
 package org.aya.core.term;
 
 import kala.collection.immutable.ImmutableSeq;
+import kala.control.Option;
+import org.aya.core.pat.PatMatcher;
 import org.jetbrains.annotations.NotNull;
 
 public record MatchTerm(
   @NotNull ImmutableSeq<Term> discriminant,
   @NotNull ImmutableSeq<Matching> clauses
 ) implements Term {
+  public @NotNull Option<Term> tryMatch() {
+    for (var clause : clauses) {
+      var subst = PatMatcher.tryBuildSubstTerms(null, clause.patterns(), discriminant.view());
+      if (subst.isOk()) {
+        return Option.some(clause.body().rename().subst(subst.get()));
+      } else if (subst.getErr()) return Option.none();
+    }
+    return Option.none();
+  }
 }
